refactor(progress-reporter-spec): Remove duplicate code
